Use relative ids
